+++ huanshankeji-team-gradle-plugins/src/main/kotlin/com/huanshankeji/team/GithubPackagesMaven.kt
@@ -1,10 +1,10 @@
 package com.huanshankeji.team
 
-import com.huanshankeji.repositoriesAddGithubPackagesMavenRegistry
+import com.huanshankeji.github.packages.repositoriesAddGithubPackagesMavenRegistry
 import org.gradle.api.Project
 
 @Deprecated(
-    "Use the context parameter version instead.", // TODO
+    "Use the context parameter version instead.",
 )
 fun Project.repositoriesAddTeamGithubPackagesMavenRegistry(repository: String) =
     repositoriesAddGithubPackagesMavenRegistry(HUANSHANKEJI_IN_LOWERCASE, repository)