+++ kotlin-common-gradle-plugins/src/main/kotlin/com/huanshankeji/github/packages/GithubPackagesMavenCredentials.kt
@@ -1,13 +1,9 @@
-package com.huanshankeji
+package com.huanshankeji.github.packages
 
+import com.huanshankeji.maven.findStringGradleProperty
+import com.huanshankeji.maven.findStringProperty
 import org.gradle.api.Project
 import org.gradle.api.initialization.Settings
-
-fun Project.findStringProperty(propertyName: String): String? =
-    findProperty(propertyName) as String?
-
-fun Settings.findStringGradleProperty(propertyName: String): String? =
-    providers.gradleProperty(propertyName).orNull
 
 fun Project.githubPackagesMavenUsername(): String? =
     findStringProperty("gpr.user") ?: findStringProperty("gprUser")
@@ -20,6 +16,3 @@ fun Settings.githubPackagesMavenUsername(): String? =
 
 fun Settings.githubPackagesMavenPassword(): String? =
     findStringGradleProperty("gpr.key") ?: findStringGradleProperty("gprKey")
-
-fun Project.gitlabPackageRegistryPrivateToken(): String? =
-    findStringProperty("gitLabPrivateToken")