+++ kotlin-common-gradle-plugins/src/main/kotlin/com/huanshankeji/gitlab/packageregistry/GitlabPackageRegistryCredentials.kt
@@ -1,0 +1,7 @@
+package com.huanshankeji.gitlab.packageregistry
+
+import com.huanshankeji.maven.findStringProperty
+import org.gradle.api.Project
+
+fun Project.gitlabPackageRegistryPrivateToken(): String? =
+    findStringProperty("gitLabPrivateToken")